requirements.txt
aiofiles==0.7.0
beautiful==0.0.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
lxml==4.6.3
mysql-connector-python==8.0.25
protobuf==3.17.0
pydantic==1.7.3
python-dotenv==0.15.0
python-multipart==0.0.5
requests==2.25.1
six==1.15.0
soap==0.0.1rc0
soupsieve==2.2.1
SQLAlchemy==1.3.23
starlette==0.13.6
urllib3==1.26.3
uvicorn==0.13.3
